Separate pin and node
